+++ get_save_params.m
@@ -1,6 +1,6 @@
 function save = get_save_params(Testmode, recobj, sobj, data)
 % save parameters for Ver11
-% these paraemters are updated in evely loop and saved in a cell.
+% these paraemters are updated in evely loop and saved in a cell-variable.
 
 global figUIobj
 
@@ -21,6 +21,7 @@ save.vbl_1 = sobj.vbl_1;
 if get(figUIobj.stim,'value') && recobj.cycleNum > 0
     save.vbl_2 = sobj.vbl_2;
     save.vbl_3 = sobj.vbl_3;
+    
     % Stimi ON %
     save.stim1.On_time = sobj.vbl_2 - sobj.vbl_1;
     save.stim1.BeamposON = sobj.BeamposON;
@@ -37,6 +38,13 @@ if get(figUIobj.stim,'value') && recobj.cycleNum > 0
     save.stim1.centerX_pix = sobj.stim_center(1);
     save.stim1.centerY_pix = sobj.stim_center(2);
     
+    %concentric position
+    if strcmp(mode, 'Concentric')
+        % concentric_position
+        save.stim1.dist_deg = sobj.concentric_mat_deg(sobj.conc_index, 1);
+        save.stim1.angle_deg = sobj.concentric_mat_deg(sobj.conc_index, 2);
+    end
+    
     %%% Size %%%
     save.stim1.size_deg = sobj.size_deg;
     save.stim1.sizeX_pix = sobj.stim_size(1);
@@ -45,10 +53,7 @@ if get(figUIobj.stim,'value') && recobj.cycleNum > 0
     %%% Stim.Pattern Specific Parameters %%%
     switch pattern
         case {'Uni', 'Size_rand'}
-            if strcmp(mode, 'Concentric')
-                save.stim1.dist_deg = sobj.concentric_mat_deg(sobj.conc_index, 1);
-                save.stim1.angle_deg = sobj.concentric_mat_deg(sobj.conc_index, 2);
-            end
+
             
         case {'2P','B/W'}
             % concentric_position stim1 or stim 2
@@ -69,11 +74,11 @@ if get(figUIobj.stim,'value') && recobj.cycleNum > 0
                 % center position
                 save.stim2.centerX_pix = sobj.stim_center2(1);
                 save.stim2.centerY_pix = sobj.stim_center2(2);
-                save.stim2.size_deg = sobj.size_deg2;
                 % size
+                save.stim2.size_deg = sobj.size_deg2;
                 save.stim2.sizeX_pix = sobj.stim_size2(1);
                 save.stim2.sizeY_pix = sobj.stim_size2(2);
-                %
+                % color luminace
                 save.stim2.lumi = sobj.lumi2;
                 save.stim2.color = sobj.stimcol2;
             end
@@ -81,28 +86,18 @@ if get(figUIobj.stim,'value') && recobj.cycleNum > 0
         case 'Looming'
             % looming speed and maximum size
             save.stim1.LoomingSpd_deg_s = sobj.loomSpd_deg;
-            save.stim1.LoomingMaxSize_deg =sobj.maxSize_deg;
-            if strcmp(mode, 'Concentric')
-                save.stim1.dist_deg = sobj.concentric_mat_deg(sobj.conc_index, 1);
-                save.stim1.angle_deg = sobj.concentric_mat_deg(sobj.conc_index, 2);
-            end
+            save.stim1.LoomingMaxSize_deg = sobj.maxSize_deg;
             
         case 'MoveBar'
             % Mogving direction and speed.
             save.stim1.MovebarSpd_deg_s = sobj.moveSpd_deg;
-            %save.stim1.MovebarDir_angle_deg = sobj.concentric_mat_deg(sobj.conc_index, 2);
-            save.stim1.MovebarDir_angle_deg =sobj.angle;
-            
+            save.stim1.MovebarDir_angle_deg = sobj.angle;
+             
         case {'Sin', 'Rect', 'Gabor'}
             % Sin, Rect, Gabor, Grating pramas
             save.stim1.gratingSF_cyc_deg = sobj.gratFreq;
             save.stim1.gratingSpd_Hz = sobj.shiftSpd;
             save.stim1.gratingAngle_deg = sobj.angle;
-            
-            if strcmp(mode, 'Concentric')
-                save.stim1.dist_deg = sobj.concentric_mat_deg(sobj.conc_index, 1);
-                save.stim1.angle_deg = sobj.concentric_mat_deg(sobj.conc_index, 2);
-            end
             
         case 'Images'
             % Image